[BEAM-13796] Move test to ReadTest class and correct javadoc for QueryTest.
